# Fix: messages stuck in mediator