--- a/aries_cloudagent/core/dispatcher.py
+++ b/aries_cloudagent/core/dispatcher.py
@@ -7,7 +7,6 @@ lifecycle hook callbacks storing state for message threads, etc.
 
 import asyncio
 import logging
-import json
 import os
 import warnings
 
@@ -23,7 +22,7 @@ from ..messaging.base_message import BaseMessage
 from ..messaging.error import MessageParseError
 from ..messaging.models.base import BaseModelError
 from ..messaging.request_context import RequestContext
-from ..messaging.responder import BaseResponder
+from ..messaging.responder import BaseResponder, SKIP_ACTIVE_CONN_CHECK_MSG_TYPES
 from ..messaging.util import datetime_now
 from ..protocols.connections.v1_0.manager import ConnectionManager
 from ..protocols.problem_report.v1_0.message import ProblemReport
@@ -377,25 +376,6 @@ class DispatcherResponder(BaseResponder):
 
         return await super().create_outbound(message, **kwargs)
 
-    async def _get_msg_type_from_enc_payload(
-        self, profile: Profile, parsed_msg: dict
-    ) -> Optional[Tuple[str, str]]:
-        """Get message type and id tuple from enc_payload."""
-        try:
-            if not isinstance(parsed_msg, dict):
-                return None
-            message_type = parsed_msg.get("@type")
-            if not message_type:
-                return None
-            registry: ProtocolRegistry = profile.inject(ProtocolRegistry)
-            message_cls = registry.resolve_message_class(message_type)
-            if not message_cls:
-                return None
-            instance = message_cls.deserialize(parsed_msg)
-            return instance._message_type, instance._id
-        except (ProtocolMinorVersionNotSupported, BaseModelError, AttributeError):
-            return None
-
     async def send_outbound(
         self, message: OutboundMessage, **kwargs
     ) -> OutboundSendStatus:
@@ -411,21 +391,14 @@ class DispatcherResponder(BaseResponder):
 
         msg_type = kwargs.get("message_type")
         msg_id = kwargs.get("message_id")
-        if not msg_type and not msg_id and (message.enc_payload or message.payload):
-            msg_dict = json.loads(message.enc_payload or message.payload)
-            msg_type_id_tuple = await self._get_msg_type_from_enc_payload(
-                context.profile, msg_dict
-            )
-            if msg_type_id_tuple:
-                msg_type, msg_id = msg_type_id_tuple
 
         if (
             message.connection_id
             and msg_type
+            and msg_type not in SKIP_ACTIVE_CONN_CHECK_MSG_TYPES
             and not await super().conn_rec_active_state_check(
                 profile=context.profile,
                 connection_id=message.connection_id,
-                msg_type=msg_type,
             )
         ):
             raise RuntimeError(