- resolve #2111

Performance Demo Analysis [No dropoff]
| branch | test config    | time taken |
|---------|--------------------------------------|-----------|
| This PR | 10000 credentials with revocation    | 1107.09 s, 1078.85 s |
| Main    | 10000 credentials with revocation    | 1089.77 s, 1075.61 s |
| This PR | 10000 credentials without revocation | 1035.22 s |
| Main    | 10000 credentials without revocation | 1042.04 s |