@ianco 2023-03-01
@dbluhm  did you check out the issue?  https://github.com/hyperledger/aries-cloudagent-python/issues/2111

@shaangill025 can answer any implementation questions