@shaangill025 2023-03-01
> @ianco yes, sorry, to be more specific, some commentary on how these changes address the issue would be helpful

The cause of the issue is sending a credential offer when the connection is not ready [in `request-sent`].

From https://github.com/fullboar/first-credential-mia#demonstration:
`sending an offer when the conneciton is in "request-sent", before it reaces a completed state, causes the offer to now show up ..`

From  https://github.com/hyperledger/aries-cloudagent-python/issues/2111:
`ACA-py as it appears to be a shortcoming in the V1 protocol where by a connection is considered valid in the "request-sent" state`

The changes will enforce that messages like cred-offer [all messages except for types included in [WHITELIST_MSG_TYPE](https://github.com/hyperledger/aries-cloudagent-python/blob/6ae3ffcd8fc0007fdfb951ff34bdfa26cda8a937/aries_cloudagent/messaging/responder.py#L139)] are only sent when a connection is ready [response or active state].

```
WHITELIST_MSG_TYPE = [
    "didexchange/1.0/request",
    "didexchange/1.0/response",
    "connections/1.0/invitation",
    "connections/1.0/request",
    "connections/1.0/response",
]
```