@ianco 2023-03-02
We need a fix because the wallet team needs an updated mediator for testing.  I wonder if it makes sense to push this update into a feature branch so that we can build a release and deploy to our own environment for testing, and then we can discuss a more "permanent" change once we confirm that this update fixes our issue.

@swcurran ?