@dbluhm 2023-03-08
I've a couple of responses. If we commit the changes @shaangill025 proposed in his comments, I think that gets this PR to a state I'm comfortable merging with the understanding that we should discuss a deeper solution :+1: