@ianco 2023-03-08
> I've left a couple of responses. If we commit the changes @shaangill025 proposed in his comments, I think that gets this PR to a state I'm comfortable merging with the understanding that we should discuss a deeper solution 👍

OK sounds good thanks!  @shaangill025 can you please do the necessary updates, and create an issue for any future work?  and then @dbluhm can contribute to the issue and we can prioritize the future work in our backlog